requirements.txt
matplotlib>=2.2.4
Pillow>=6.0.0
scikit-learn>=0.20.3
scipy>=1.2.1
tensorflow>=1.13.1,<2.0

